+++ Final2/src/edu/kit/informatik/wtrs/time/WorkTime.java
@@ -1,7 +1,5 @@
 package edu.kit.informatik.wtrs.time;
 
-import edu.kit.informatik.wtrs.ui.Main;
-
 import java.util.ArrayList;
 import java.util.Collection;
 
@@ -20,21 +18,21 @@ public class WorkTime extends Interval implements Comparable<Interval> {
         this(workTimeStart, workTimeEnd, null, null);
     }
 
-    public int pauseInMinutes() {
+    public long pauseInMinutes() {
         return this.pause.duration().toMinutes();
     }
 
-    public int pureWorkInMinutes() {
+    public long pureWorkInMinutes() {
         return Duration.differenceBetween(this.duration(), this.pause.duration()).toMinutes();
     }
 
     @Override
-    public int pureDurationInMinutesBefore(Date date) {
+    public long pureDurationInMinutesBefore(Date date) {
         return this.minutesBefore(date) - this.pause.minutesBefore(date);
     }
 
-    public Collection<Integer> workTimeBlocksInMinutes() {
-        Collection<Integer> durations = new ArrayList<>();
+    public Collection<Long> workTimeBlocksInMinutes() {
+        Collection<Long> durations = new ArrayList<>();
 
         durations.add(this.getStart().durationTo(this.pause.getStart()).toMinutes());
         durations.add(this.pause.getEnd().durationTo(this.getEnd()).toMinutes());
